requirements.txt
pyyaml>=6.0
requests>=2.28.0
jsonschema>=4.23.0